Mobile replacement for Special:Watchlist, initial version.
So far includes a simplified overview display, but not yet a drilldown into individual changes.

A basic filter is available to select between 'all' pages, article pages, article talk pages, and other pages -- this needs further work as well.

May need to consider more issues of tagging/marking.

Note that some of SpecialMobileWatchlist's code is copied from core SpecialWatchlist, with the intention that it will be refactored further into a base that works with both desktop and mobile.

Change-Id: Iee1a4bfde4369df0f1f0b1dba3a65d45219e8efa