Detect disabled file upload controls on Firefox OS 1.0
Firefox for Android and desktop support file uploads, but the actual
<input type="file"> is not yet supported on Firefox OS. As of 1.0, it
silently drops the "type='file'" and turns file upload controls into
text fields.

This is kind of annoying, as otherwise all the infrastructure is there,
such as FormData and FileReader.

Patch adds detection for the Firefox OS case which will gracefully
start picking up the functioning controls when they get added in future
versions.

Change-Id: I9bfefa39b56c97b65116c9e1f18409cb10d06867